adding logging to help with debug

+++ controllers/tests/kafkacluster_controller_test.go
@@ -549,8 +549,15 @@ var _ = Describe("KafkaCluster with two config external listener", func() {
 		// SafeKafkaCleanup(ctx, k8sClient, kafkaCluster, kafkaClusterKRaft, namespace)
 
 		By("deleting Kafka cluster object " + kafkaCluster.Name + " in namespace " + namespace)
-		err := k8sClient.Delete(ctx, kafkaCluster)
-		Expect(err).NotTo(HaveOccurred())
+
+		if kafkaCluster != nil {
+			fmt.Printf(" Context before delete: ", ctx)
+			fmt.Printf(" kafkacluster: ", kafkaCluster)
+			err := k8sClient.Delete(ctx, kafkaCluster)
+			Expect(err).NotTo(HaveOccurred())
+		} else {
+			fmt.Printf(" kafkacluster Already Nil!")
+		}
 
 		kafkaCluster = nil
 		kafkaClusterKRaft = nil
